+++ SmartWallet/Models/CashVault/UpdateTransactionEndpointApiModel.cs
@@ -18,5 +18,5 @@ public class UpdateTransactionEndpointApiModel
 	/// <summary>
 	/// Ограничение трат
 	/// </summary>
-	public double? Limitation { get; set; }
+	public decimal? Limitation { get; set; }
 }
