+++ SmartWallet/Models/FinancialAnalytics/CategoryComparativeAnalysisResponse.cs
@@ -8,12 +8,12 @@ public sealed class CategoryComparativeAnalysisResponse
     /// <summary>
     /// Общая сумма трат во втором периоде
     /// </summary>
-    public double TotalSecondPeriodSpending { get; set; }
+    public decimal TotalSecondPeriodSpending { get; set; }
     
     /// <summary>
     /// Общая сумма трат в первом периоде
     /// </summary>
-    public double TotalFirstPeriodSpending { get; set; }
+    public decimal TotalFirstPeriodSpending { get; set; }
 
     /// <summary>
     /// Сравнительные анализы по категориям
